0.3.0
batches now work correctly, and fixed validation accuracy bug and status bar bug. And added eta stat.

+++ NeuralNetViewer/Widgets/TrainWidgets/trainstatuswidget.cpp
@@ -46,7 +46,7 @@ TrainStatusWidget::TrainStatusWidget(QWidget* parent)
 }
 
 void TrainStatusWidget::setEpochTrainingData(const EpochStats& epoch_stats){
-    m_status_bar->setProgress(epoch_stats.epoch / epoch_stats.total_epochs);
+    m_status_bar->setProgress(static_cast<float>(epoch_stats.epoch) / static_cast<float>(epoch_stats.total_epochs));
     m_stats_frame->set_epoch_stats(epoch_stats);
     m_graph_widget->update_training_loss_data(epoch_stats.epoch, epoch_stats.training_loss);
     m_graph_widget->update_validation_loss_data(epoch_stats.epoch, epoch_stats.training_loss);
